transitioning functionality to handlers
